// packages/typespec-ts/src/rlc-common/interfaces.ts
export type Methods = {
  // could be more than one method if overloading
  [key: string]: OperationMethod[];
};

export interface ResponseTypes {
  success: string[];
  error: string[];
}

export interface OperationMethod {
  optionsName: string;
  description: string;
  hasOptionalOptions: boolean;
  returnType: string;
  successStatus: string[];
  responseTypes: ResponseTypes;
  operationName: string;
  operationHelperDetail?: OperationHelperDetail;
}
export interface PathMetadata {
  name: string;
  pathParameters: PathParameter[];
  methods: Methods;
  operationGroupName: string;
  description: string;
}

export type Paths = Record<string, PathMetadata>;

export type PathParameter = {
  oriName?: string;
  name: string;
  documentName?: string;
  type: string;
  description?: string;
  value?: string | number | boolean;
  wrapperType?: Schema;
};
// ...
export enum SchemaContext {
  /** Schema is used as an input to an operation. */
  Input = "input",
  /** Schema is used as an output from an operation. */
  Output = "output",
  /** Schema is used as an exception from an operation. */
  Exception = "exception"
}

export interface Schema {
  name: string;
  type: string;
  typeName?: string;
  outputTypeName?: string;
  description?: string;
  required?: boolean;
  default?: any;
  readOnly?: boolean;
  usage?: SchemaContext[];
  alias?: string;
  outputAlias?: string;
  fromCore?: boolean;
  enum?: any[];
  isConstant?: boolean;
}